refactor: ProjectService 리팩토링 및 코드 구조 개선
1. 클래스 패키지 변경 및 의존성 추가/변경을 수행한다.
2. 프로젝트 생성 로직을 Factory 패턴으로 리팩토링한다.
3. 불필요한 private 메서드 및 update 관련 로직을 제거한다.

Resolves #28
